Se generan las UT existentes y se recuperan otras previas, se guarda todo en Test.
