+++ Test/resources/cancelByXml.xml
