+++ Test/resources/fileAcceptReject.xml
