+++ Test/resources/filePdf.xml
@@ -1,0 +1,35 @@
+<?xml version= "1.0" encoding= "utf-8"?>
+<cfdi:Comprobante xsi:schemaLocation= "http://www.sat.gob.mx/cfd/4 http://www.sat.gob.mx/sitio_internet/cfd/4/cfdv40.xsd" Version= "4.0" Serie= "Serie" Folio= "Folio" Fecha= "2023-05-29T00:00:55" Sello= "ed+Xadb7aJHsu7nPMVjPjKOWUPewkDrN4Lv5XC8geHB7GYPshexjm+k6+M7FVBZUFFK5E+Aug5sYd7w9kn/DzRCLijyJhOz7LvleCyEGFUiLt9AQqRoYk0iQBcik+Bzz1gfz2qwelPEhODnb/gVquckkdt1woqW0/Gh9xfISY3ipnpgh1epaxP4MDCCWYX66XdH2vTwx4fANZTh7XYoXpkkQpoXT0xEuJQDyKiQnUMxiijBO/3EyDDc8Vh+vTBy5ev0lVM62Fe9Q4MoYDwdG9rY7Kph+whTOHXCkJ2UfH4WtsFlIDEoXNGtXXL6fWUFhA+lzv4vtQwtkb4LQFqoT3w=="
+    CondicionesDePago= "CondicionesDePago" SubTotal= "200" Descuento= "1" Moneda= "AMD" TipoCambio= "1" Total= "198.96"
+    TipoDeComprobante= "I" Exportacion= "01" MetodoPago= "PPD" FormaPago= "99" LugarExpedicion= "20000" xmlns:cfdi= "http://www.sat.gob.mx/cfd/4"
+    xmlns:xsi= "http://www.w3.org/2001/XMLSchema-instance" NoCertificado= "30001000000400002434" Certificado= "MIIFuzCCA6OgAwIBAgIUMzAwMDEwMDAwMDA0MDAwMDI0MzQwDQYJKoZIhvcNAQELBQAwggErMQ8wDQYDVQQDDAZBQyBVQVQxLjAsBgNVBAoMJVNFUlZJQ0lPIERFIEFETUlOSVNUUkFDSU9OIFRSSUJVVEFSSUExGjAYBgNVBAsMEVNBVC1JRVMgQXV0aG9yaXR5MSgwJgYJKoZIhvcNAQkBFhlvc2Nhci5tYXJ0aW5lekBzYXQuZ29iLm14MR0wGwYDVQQJDBQzcmEgY2VycmFkYSBkZSBjYWRpejEOMAwGA1UEEQwFMDYzNzAxCzAJBgNVBAYTAk1YMRkwFwYDVQQIDBBDSVVEQUQgREUgTUVYSUNPMREwDwYDVQQHDAhDT1lPQUNBTjERMA8GA1UELRMIMi41LjQuNDUxJTAjBgkqhkiG9w0BCQITFnJlc3BvbnNhYmxlOiBBQ0RNQS1TQVQwHhcNMTkwNjE3MTk0NDE0WhcNMjMwNjE3MTk0NDE0WjCB4jEnMCUGA1UEAxMeRVNDVUVMQSBLRU1QRVIgVVJHQVRFIFNBIERFIENWMScwJQYDVQQpEx5FU0NVRUxBIEtFTVBFUiBVUkdBVEUgU0EgREUgQ1YxJzAlBgNVBAoTHkVTQ1VFTEEgS0VNUEVSIFVSR0FURSBTQSBERSBDVjElMCMGA1UELRMcRUtVOTAwMzE3M0M5IC8gWElRQjg5MTExNlFFNDEeMBwGA1UEBRMVIC8gWElRQjg5MTExNk1HUk1aUjA1MR4wHAYDVQQLExVFc2N1ZWxhIEtlbXBlciBVcmdhdGUwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCN0peKpgfOL75iYRv1fqq+oVYsLPVUR/GibYmGKc9InHFy5lYF6OTYjnIIvmkOdRobbGlCUxORX/tLsl8Ya9gm6Yo7hHnODRBIDup3GISFzB/96R9K/MzYQOcscMIoBDARaycnLvy7FlMvO7/rlVnsSARxZRO8Kz8Zkksj2zpeYpjZIya/369+oGqQk1cTRkHo59JvJ4Tfbk/3iIyf4H/Ini9nBe9cYWo0MnKob7DDt/vsdi5tA8mMtA953LapNyCZIDCRQQlUGNgDqY9/8F5mUvVgkcczsIgGdvf9vMQPSf3jjCiKj7j6ucxl1+FwJWmbvgNmiaUR/0q4m2rm78lFAgMBAAGjHTAbMAwGA1UdEwEB/wQCMAAwCwYDVR0PBAQDAgbAMA0GCSqGSIb3DQEBCwUAA4ICAQBcpj1TjT4jiinIujIdAlFzE6kRwYJCnDG08zSp4kSnShjxADGEXH2chehKMV0FY7c4njA5eDGdA/G2OCTPvF5rpeCZP5Dw504RZkYDl2suRz+wa1sNBVpbnBJEK0fQcN3IftBwsgNFdFhUtCyw3lus1SSJbPxjLHS6FcZZ51YSeIfcNXOAuTqdimusaXq15GrSrCOkM6n2jfj2sMJYM2HXaXJ6rGTEgYmhYdwxWtil6RfZB+fGQ/H9I9WLnl4KTZUS6C9+NLHh4FPDhSk19fpS2S/56aqgFoGAkXAYt9Fy5ECaPcULIfJ1DEbsXKyRdCv3JY89+0MNkOdaDnsemS2o5Gl08zI4iYtt3L40gAZ60NPh31kVLnYNsmvfNxYyKp+AeJtDHyW9w7ftM0Hoi+BuRmcAQSKFV3pk8j51la+jrRBrAUv8blbRcQ5BiZUwJzHFEKIwTsRGoRyEx96sNnB03n6GTwjIGz92SmLdNl95r9rkvp+2m4S6q1lPuXaFg7DGBrXWC8iyqeWE2iobdwIIuXPTMVqQb12m1dAkJVRO5NdHnP/MpqOvOgLqoZBNHGyBg4Gqm4sCJHCxA1c8Elfa2RQTCk0tAzllL4vOnI1GHkGJn65xokGsaU4B4D36xh7eWrfj4/pgWHmtoDAYa8wzSwo2GVCZOs+mtEgOQB91/g==">
+    <cfdi:Emisor Rfc= "EKU9003173C9" Nombre= "ESCUELA KEMPER URGATE" RegimenFiscal= "601" />
+    <cfdi:Receptor Rfc= "URE180429TM6" Nombre= "UNIVERSIDAD ROBOTICA ESPAÑOLA" DomicilioFiscalReceptor= "65000" RegimenFiscalReceptor= "601" UsoCFDI= "G01" />
+    <cfdi:Conceptos>
+        <cfdi:Concepto ClaveProdServ= "50211503" Cantidad= "1" ClaveUnidad= "H87" Unidad= "Pieza" Descripcion= "Cigarros"
+            ValorUnitario= "200.00" Descuento= "1" Importe= "200.00" ObjetoImp= "02">
+            <cfdi:Impuestos>
+                <cfdi:Traslados>
+                    <cfdi:Traslado Base= "1" Importe= "0.16" Impuesto= "002" TasaOCuota= "0.160000" TipoFactor= "Tasa" />
+                </cfdi:Traslados>
+                <cfdi:Retenciones>
+                    <cfdi:Retencion Base= "1" Impuesto= "001" TipoFactor= "Tasa" TasaOCuota= "0.100000" Importe= "0.10" />
+                    <cfdi:Retencion Base= "1" Impuesto= "002" TipoFactor= "Tasa" TasaOCuota= "0.106666" Importe= "0.10" />
+                </cfdi:Retenciones>
+            </cfdi:Impuestos>
+        </cfdi:Concepto>
+    </cfdi:Conceptos>
+    <cfdi:Impuestos TotalImpuestosRetenidos= "0.20" TotalImpuestosTrasladados= "0.16">
+        <cfdi:Retenciones>
+            <cfdi:Retencion Impuesto= "001" Importe= "0.10" />
+            <cfdi:Retencion Impuesto= "002" Importe= "0.10" />
+        </cfdi:Retenciones>
+        <cfdi:Traslados>
+            <cfdi:Traslado Base= "1" Importe= "0.16" Impuesto= "002" TasaOCuota= "0.160000" TipoFactor= "Tasa" />
+        </cfdi:Traslados>
+    </cfdi:Impuestos>
+    <cfdi:Complemento>
+        <tfd:TimbreFiscalDigital xsi:schemaLocation= "http://www.sat.gob.mx/TimbreFiscalDigital http://www.sat.gob.mx/sitio_internet/cfd/TimbreFiscalDigital/TimbreFiscalDigitalv11.xsd" Version= "1.1" UUID= "2dfa4bb4-eae4-4a9b-b69d-526a90092a31" FechaTimbrado= "2023-05-29T16:43:10"
+            RfcProvCertif= "SPR190613I52" SelloCFD= "ed+Xadb7aJHsu7nPMVjPjKOWUPewkDrN4Lv5XC8geHB7GYPshexjm+k6+M7FVBZUFFK5E+Aug5sYd7w9kn/DzRCLijyJhOz7LvleCyEGFUiLt9AQqRoYk0iQBcik+Bzz1gfz2qwelPEhODnb/gVquckkdt1woqW0/Gh9xfISY3ipnpgh1epaxP4MDCCWYX66XdH2vTwx4fANZTh7XYoXpkkQpoXT0xEuJQDyKiQnUMxiijBO/3EyDDc8Vh+vTBy5ev0lVM62Fe9Q4MoYDwdG9rY7Kph+whTOHXCkJ2UfH4WtsFlIDEoXNGtXXL6fWUFhA+lzv4vtQwtkb4LQFqoT3w==" NoCertificadoSAT= "30001000000400002495" SelloSAT= "omtsaiXiOJbyLiSITJEroUOwzDWhyrVfeZEtzRv1MTj1AtulefsnK98JiupkVyIckj++oGxPYFO94lAxdkqcjtlDZZ5ylZ4gRrDZqptqahSzTF/ZUAY+deo903Aww+GTaP0jjFtKIliShjOUHLzHbMNtDpx2M9/v9LfBjIBNGPylhzp3ZuXpECtLywS3bOyfndITGc+xloMVtqxWDx7yvoVuAhykzEjWxm5Qzj7aYn7H03Xlb1NlclSHdrGkLzubkm7OHV3iIqKRhDiRAwa0eZVUZcfxEQ+RvueHZ9+SY9S22gJWa13QAmPsG1eWHVc4GgBeGO9tdSrYQhg7O80+3w==" xmlns:tfd= "http://www.sat.gob.mx/TimbreFiscalDigital" xmlns:xsi= "http://www.w3.org/2001/XMLSchema-instance" />
+    </cfdi:Complemento>
+</cfdi:Comprobante>
