+++ Test/test_validate.py
@@ -1,0 +1,24 @@
+import unittest
+import os
+import json
+import sys
+
+PROJECT_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir))
+sys.path.append(PROJECT_ROOT)
+
+from Validate.Validate import Validate
+
+class MyTest(unittest.TestCase):
+    expected = "success"
+    message = "307. El comprobante contiene un timbre previo."
+    @staticmethod
+    def open_file(pathFile):
+        out = open(pathFile, "r", encoding='ansi', errors='ignore').read()
+        return out
+    def testValidateXml(self):
+        validate = Validate("http://services.test.sw.com.mx", None, "david.reyes@sw.com.mx", "TEST2022")
+        response = validate.ValidateXml(MyTest.open_file("resources\\xml40.xml"))
+        self.assertTrue(self.expected == response.get_status())
+
+suite = unittest.TestLoader().loadTestsFromTestCase(MyTest)
+unittest.TextTestRunner(verbosity=2).run(suite)
